backend/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
SQLAlchemy==2.0.30
Werkzeug==3.0.1
python-dotenv==1.0.1
Flask-Migrate==4.0.5
flask_jwt_extended==4.6.0
gunicorn==21.2.0
azure-identity==1.15.0
azure-core==1.30.0
pandas==2.2.3
